fix set tests

+++ tests/table_test.c
@@ -26,15 +26,15 @@ void
 basic_tests(struct zt_unit_test *test, void *data){
 
 	char	  	  buf[1024];
-	int 		  i;
-	int 		  nn = 0;
-	int		  	  numbers[10] = { 9,8,7,6,5,4,3,2,1,0 };
+	ssize_t		  i;
+	ssize_t		  nn = 0;
+	ssize_t	  	  numbers[10] = { 9,8,7,6,5,4,3,2,1,0 };
 	zt_table	* table_int = NULL;
 	zt_table 	* table_str = NULL;
   
 	table_int = zt_table_init("test int table", zt_table_hash_int, zt_table_compare_int, 10, 0, NULL);
 	table_str = zt_table_init("test string table", zt_table_hash_string,
-						   zt_table_compare_string, 10, 0, NULL);
+							   zt_table_compare_string, 10, 0, NULL);
   
 	for( i=9; i >= 0; i--)
 		zt_table_set (table_int, (void *)i, (void *)numbers[i]);
@@ -43,7 +43,7 @@ basic_tests(struct zt_unit_test *test, void *data){
 
 	ZT_UNIT_ASSERT(test, count == 10);
 	
-	nn = (int)zt_table_get ( table_int, (void *)5 );
+	nn = (ssize_t)zt_table_get ( table_int, (void *)5 );
 	ZT_UNIT_ASSERT(test, nn == 4);
 
 	/* string test */
@@ -77,8 +77,8 @@ str_free(void *key, void *datum, void *param)
 static int
 str_iterator (void *key, void *datum, void *param)
 {
-	int 				  d = (int)datum;	
-	int 				  k = atoi((key + strlen(STR_TEST_PRE)));
+	ssize_t 			  d = (intptr_t)datum;	
+	ssize_t 			  k = atoi((key + strlen(STR_TEST_PRE)));
 	struct zt_unit_test	* test = (struct zt_unit_test *)param;
 
 	ZT_UNIT_ASSERT(test, (k + d == 9));
@@ -90,8 +90,8 @@ str_iterator (void *key, void *datum, void *param)
 static int
 int_iterator (void *key, void *datum, void *param)
 {
-	int					  d = (int)datum;
-	int					  k = (int)key;
+	ssize_t				  d = (intptr_t)datum;
+	ssize_t				  k = (intptr_t)key;
 	struct zt_unit_test	* test = (struct zt_unit_test *)param;
 	
 	ZT_UNIT_ASSERT(test, (k + d == 9));
